feat(core-flows,medusa): remove cart customer validation + rename workflow
