+++ .changeset/odd-dolls-tie.md
@@ -1,0 +1,6 @@
+---
+"@medusajs/core-flows": patch
+"@medusajs/medusa": patch
+---
+
+feat(core-flows,medusa): remove cart customer validation + rename workflow
