+++ packages/core/core-flows/src/cart/workflows/index.ts
@@ -6,8 +6,8 @@ export * from "./create-carts"
 export * from "./create-payment-collection-for-cart"
 export * from "./list-shipping-options-for-cart"
 export * from "./refresh-payment-collection"
+export * from "./transfer-cart-customer"
 export * from "./update-cart"
-export * from "./update-cart-customer"
 export * from "./update-cart-promotions"
 export * from "./update-line-item-in-cart"
 export * from "./update-tax-lines"
